Rename servweb subpackge to clientweb
- abort unused subpackages servccdb and servname and delete their file.
