+++ siriuspy/siriuspy/__init__.py
@@ -4,9 +4,10 @@ with open(_os.path.join(__path__[0], 'VERSION'), 'r') as _f:
 del _os
 
 
-__all__ = ['envars', 'util', 'servname', 'servccdb', 'servweb', 'servconf',
-           'diagnostics', 'pwrsupply', 'magnet', 'namesys', 'timesys',
-           'csdevice', 'epics', 'callbacks']
+__all__ = [
+    'envars', 'util', 'clientweb', 'clientconfigdb', 'clientarch'
+    'diagnostics', 'pwrsupply', 'magnet', 'namesys', 'timesys', 'csdevice',
+    'epics', 'callbacks']
 
 
 subpackages = __all__.copy()
