--- siriuspy/siriuspy/servname/__init__.py
@@ -1,3 +1,0 @@
-from .implementation import *
-
-del implementation
